requirements.txt
crewai>=0.1.32
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-community>=0.0.10
beautifulsoup4>=4.12.2
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.2
selenium>=4.15.2
duckduckgo-search>=3.9.9
openai>=1.12.0
lxml>=4.9.3
python-docx>=0.8.11
openpyxl>=3.1.2
pandas>=2.0.0